Added support for play count, last play time and playback position..
- Play count is incremented whenever Playback calls onPlaybackComplete(). The lastPlayed date is also recorded at this time.
- Playback position is now stored in the database, so that it can be restored at a later date. This occurs when the track is paused, or when playback completes.
- Songs now have a 'type' property, which is either 'normal', 'podcast', or 'audiobook'. Type is determined based on the file path and extension.
- A new load() method exists in the PlaybackManager, which will restore the seek position for audiobooks and podcasts before playback.
- Song position and duration is now stored throughout the app as integer instead of long, as there's no real need to store it as a long.
- onPlaybackComplete() now returns a non-optional Song object, since playback can't really complete if the song that was playing is null.

+++ app/src/main/java/com/simplecityapps/shuttle/ui/common/StringExt.kt
@@ -2,10 +2,10 @@ package com.simplecityapps.shuttle.ui.common
 
 import java.util.concurrent.TimeUnit
 
-fun Long.toHms(): String {
-    val hours = TimeUnit.MILLISECONDS.toHours(this)
-    val minutes = TimeUnit.MILLISECONDS.toMinutes(this) % TimeUnit.HOURS.toMinutes(1)
-    val seconds = TimeUnit.MILLISECONDS.toSeconds(this) % TimeUnit.MINUTES.toSeconds(1)
+fun Int.toHms(): String {
+    val hours = TimeUnit.MILLISECONDS.toHours(this.toLong())
+    val minutes = TimeUnit.MILLISECONDS.toMinutes(this.toLong()) % TimeUnit.HOURS.toMinutes(1)
+    val seconds = TimeUnit.MILLISECONDS.toSeconds(this.toLong()) % TimeUnit.MINUTES.toSeconds(1)
 
     return if (hours == 0L) {
         String.format("%2d:%02d", minutes, seconds)
